# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "creationGoal-api",
    "description": "CreationGoal Process App Personal Project",
    "version": "1.0.0"
  },
  "host": "p2-web-services-w4.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Registers a user ---(AUTH DOORWAY FOR PASSWORD SIGN IN)---",
        "description": "Special route created for posting the registration of new users for password sign-in capability.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Fields to fill out.   \n(Additional OPTIONAL FIELDS you can add to the BODY BELOW = 'creationNumber':, 'creationDate':, & 'status':)",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "email@email.com"
                },
                "password": {
                  "type": "password",
                  "example": "password123"
                },
                "repeat password": {
                  "type": "string",
                  "example": "password123"
                },
                "required": "email"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "summary": "DELETES ---(OAUTH AUTHORIZATION)--- for the user",
        "description": "After DELETING AUTHORIZATION of the user it returns a success code and redirects the user to the LOGIN PAGE.",
        "responses": {
          "200": {
            "description": "SUCCESS, the OAUTH AUTORIZATION was DELETED"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/dashboard/": {
      "get": {
        "summary": "GETS all creationGoals associated with the _id of a user - essentially the LOGGED IN HOME ---------- (!!!OAUTH PROTECTED ROUTE!!!)",
        "description": "All creationGoals associated with that user are displayed on the dashboard page.",
        "parameters": [
          {
            "name": "registered",
            "in": "query",
            "description": "(true) indicates a new user was successfully created, this trigger a post to the dashboard page that the user is redirected to.",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "created",
            "in": "query",
            "description": "(true) indicates a new creationGoal was successfully created, this trigger a post to the dashboard page that the user is redirected to.",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "updated",
            "in": "query",
            "description": "(true) indicates a new creationGoal was successfully updated, this trigger a post to the dashboard page that the user is redirected to.",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "deleted",
            "in": "query",
            "description": "(true) indicates a new creationGoal was successfully deleted, this trigger a post to the dashboard page that the user is redirected to.",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "SUCCESS, GET returned all creationGoals associated with the user, after a successful loading of page.   \nRedirection SUCCESS loading options: #1 after a CREATOINGOAL was UPDATED, #2 after a CREATOINGOAL was DELETED"
          },
          "201": {
            "description": "SUCCESS, GET returned all creationGoals associated with the user, after ADDING another CREATIONGOAL"
          },
          "401": {
            "description": "You are NOT AUTHORIZED to GET the creationGoals"
          },
          "500": {
            "description": "There was an INTERNAL SERVER ERROR while trying to GET all creationGoals associated with the user"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}